+++ app/dance/spinners/triangle.go
@@ -10,21 +10,21 @@ import (
 var indicesTriangle = []mgl32.Vec3{{-0.86602540378, -0.5, 0}, {0.86602540378, -0.5, 0}, {0, 1, 0}}
 
 type TriangleMover struct {
-	start int64
+	start float64
 }
 
 func NewTriangleMover() *TriangleMover {
 	return &TriangleMover{}
 }
 
-func (c *TriangleMover) Init(start, end int64) {
+func (c *TriangleMover) Init(start, end float64) {
 	c.start = start
 }
 
-func (c *TriangleMover) GetPositionAt(time int64) vector.Vector2f {
+func (c *TriangleMover) GetPositionAt(time float64) vector.Vector2f {
 	mat := mgl32.Rotate3DZ(float32(time-c.start) / 2000 * 2 * math32.Pi).Mul3(mgl32.Scale2D(float32(settings.Dance.SpinnerRadius), float32(settings.Dance.SpinnerRadius)))
 
-	startIndex := ((time - c.start) / 10) % 3
+	startIndex := (int64(time - c.start) / 10) % 3
 
 	pt1 := indicesTriangle[startIndex]
 
@@ -36,7 +36,7 @@ func (c *TriangleMover) GetPositionAt(time int64) vector.Vector2f {
 	pt1 = mat.Mul3x1(pt1)
 	pt2 = mat.Mul3x1(pt2)
 
-	t := float32((time-c.start)%10) / 10
+	t := float32(int64(time-c.start)%10) / 10
 
 	return vector.NewVec2f((pt2.X()-pt1.X())*t+pt1.X(), (pt2.Y()-pt1.Y())*t+pt1.Y()).Add(center)
 }
